cleans up main layout page
